### 2024-03-20 d247e3b
update doc requirements
