+++ docs/_sources/auto_examples/03_eddy-lifetime_fit_3term_loss.ipynb.txt
@@ -132,7 +132,7 @@
       "name": "python",
       "nbconvert_exporter": "python",
       "pygments_lexer": "ipython3",
-      "version": "3.9.13"
+      "version": "3.9.16"
     }
   },
   "nbformat": 4,
