+++ docs/_sources/auto_examples/index.rst.txt
@@ -19,17 +19,15 @@ The first two examples correspond to the experiments presented in `the original 
 
     04_eddy-lifetime_fit_gelu.py
 
-    05_custorm_noisy_data_fit.py
+    05_custom_noisy_data_fit.py
 
-    06_custorm_data_interpolate_and_fit.py
+    06_custom_data_interpolate_and_fit.py
 
     07_mann_linear_regression.py
 
     08_mann_box_generation_IEC.py
 
     09_drd_box.py
-
-    
 
 
 
