+++ docs/_sources/auto_examples/04_eddy-lifetime_fit_gelu.ipynb.txt
@@ -150,7 +150,7 @@
       "name": "python",
       "nbconvert_exporter": "python",
       "pygments_lexer": "ipython3",
-      "version": "3.9.13"
+      "version": "3.9.16"
     }
   },
   "nbformat": 4,
