+++ docs/_sources/auto_examples/09_drd_box.ipynb.txt
@@ -157,7 +157,7 @@
       "name": "python",
       "nbconvert_exporter": "python",
       "pygments_lexer": "ipython3",
-      "version": "3.9.13"
+      "version": "3.9.16"
     }
   },
   "nbformat": 4,
