+++ docs/_sources/auto_examples/02_eddy-lifetime_fit.ipynb.txt
@@ -204,7 +204,7 @@
       "name": "python",
       "nbconvert_exporter": "python",
       "pygments_lexer": "ipython3",
-      "version": "3.9.13"
+      "version": "3.9.16"
     }
   },
   "nbformat": 4,
