### 2024-04-01 851d380
:arrow_up: Bump torch from 1.13.0 to 2.2.2
Bumps [torch](https://github.com/pytorch/pytorch) from 1.13.0 to 2.2.2.
- [Release notes](https://github.com/pytorch/pytorch/releases)
- [Changelog](https://github.com/pytorch/pytorch/blob/main/RELEASE.md)
- [Commits](https://github.com/pytorch/pytorch/compare/v1.13.0...v2.2.2)

---
updated-dependencies:
- dependency-name: torch
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
