### 2024-04-01 1c12b30
Update paper.md with NREL edits
