+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'DRDMannTurb: A python package for scalable, data-driven synthetic turbulence'
+title: 'DRDMannTurb: A Python package for scalable, data-driven synthetic turbulence'
 tags:
   - Python
   - Torch
@@ -26,13 +26,13 @@ bibliography: paper.bib
 
 # Summary
 
-Synthetic turbulence models (STMs) are used in wind engineering to generate realistic flow fields and employed as inputs to wind simulations. Examples include prescribing inlet conditions and inflow wind fields to model loads on wind turbines and tall buildings. We are interested in STMs capable of generating fluctuations based on prescribed second-moment statistics as this allows for simulating environmental conditions closely resembling on-site observations. To this end, the widely used Mann model [@mann_spatial:1994; @mann_wind:1998] is the inspiration for `DRDMannTurb`. Three physical parameters describe turbulence in this model: a magnitude parameter corresponding to the Kolmogorov constant multiplied by the rate of viscous dissipation of the turbulent kinetic energy to the two thirds, $\alpha \epsilon^{2/3}$, a turbulence length scale parameter $L$, and a non-dimensional parameter $\Gamma$ related to the lifetime of the eddies. A number of studies as well as international standards (e.g., the International Electrotechnical Commission) include recommendations for values of these three parameters with the goal of standardizing wind simulations to realistic energy spectra. Yet, having only three parameters, the Mann model faces limitations in accurately representing the diversity of observable spectra. This Python package allows users to extend the Mann model and more accurately fit field measurements through a neural network model of the eddy lifetime function. Following [@Keith:2021], we refer to this class of models as deep rapid distortion (DRD) models. Once calibrated, a DRD model is no more expensive for simulating turbulence than a Mann model. Moreover, `DRDMannTurb` comes with a state-of-the-art domain decomposition-based simulator for generating turbulence boxes for both Mann and DRD models.
+Synthetic turbulence models (STMs) are used in wind engineering to generate realistic flow fields and employed as inputs to wind simulations. Examples include prescribing inlet conditions and inflow wind fields to model loads on wind turbines and tall buildings. We are interested in STMs capable of generating fluctuations based on prescribed second-moment statistics, as this allows for simulating environmental conditions closely resembling on-site observations. To this end, the widely used Mann model [@mann_spatial:1994; @mann_wind:1998] is the inspiration for `DRDMannTurb`. Three physical parameters describe turbulence in this model: a magnitude parameter corresponding to the Kolmogorov constant multiplied by the rate of viscous dissipation of the turbulent kinetic energy to the two thirds, $\alpha \epsilon^{2/3}$, a turbulence length scale parameter $L$, and a non-dimensional parameter $\Gamma$ related to the lifetime of the eddies. A number of studies as well as international standards (e.g., the International Electrotechnical Commission) include recommendations for values of these three parameters with the goal of standardizing wind simulations to realistic energy spectra. Yet, having only three parameters, the Mann model faces limitations in accurately representing the diversity of observable spectra. This Python package allows users to extend the Mann model and more accurately fit field measurements through a neural network model of the eddy lifetime function. Following [@Keith:2021], we refer to this class of models as deep rapid distortion (DRD) models. Once calibrated, a DRD model is no more expensive for simulating turbulence than a Mann model. Moreover, `DRDMannTurb` comes with a state-of-the-art domain decomposition-based simulator for generating turbulence boxes for both Mann and DRD models.
 
 # Statement of need
 
[…]
