+++ requirements.txt
@@ -5,7 +5,7 @@ pyevtk==1.6.0
 pyFFTW==0.13.1
 pytest==7.4.2
 scikit_learn==1.3.0
-scipy==1.10.1
+scipy==1.13.0
 setuptools==67.8.0
 sphinx_gallery==0.15.0
 tensorboard==2.15.0
