### 2024-04-08 c6c6210
:arrow_up: Bump matplotlib from 3.7.1 to 3.8.4
Bumps [matplotlib](https://github.com/matplotlib/matplotlib) from 3.7.1 to 3.8.4.
- [Release notes](https://github.com/matplotlib/matplotlib/releases)
- [Commits](https://github.com/matplotlib/matplotlib/compare/v3.7.1...v3.8.4)

---
updated-dependencies:
- dependency-name: matplotlib
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
