+++ requirements.txt
@@ -1,4 +1,4 @@
-matplotlib==3.7.1
+matplotlib==3.8.4
 numpy==1.24.3
 plotly==5.17.0
 pyevtk==1.6.0
