### 2024-04-15 f447da6
:arrow_up: Bump scikit-learn from 1.3.0 to 1.4.2
Bumps [scikit-learn](https://github.com/scikit-learn/scikit-learn) from 1.3.0 to 1.4.2.
- [Release notes](https://github.com/scikit-learn/scikit-learn/releases)
- [Commits](https://github.com/scikit-learn/scikit-learn/compare/1.3.0...1.4.2)

---
updated-dependencies:
- dependency-name: scikit-learn
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
