+++ requirements.txt
@@ -4,7 +4,7 @@ plotly==5.17.0
 pyevtk==1.6.0
 pyFFTW==0.13.1
 pytest==7.4.2
-scikit_learn==1.3.0
+scikit_learn==1.4.2
 scipy==1.10.1
 setuptools==67.8.0
 sphinx_gallery==0.15.0
