### 2024-04-22 #6506 @crvernon
@jbytecode I know this one may be a little out of area of interest a little, but would you mind taking this one on as editor?
