### 2024-04-23 #6506 @HaoZeke
👋 @alizma I will facilitate the review of your paper/software as the editor. The first step will be to find reviewers, so I will start contacting potential reviewers over the weekend.

If you have any suggestions for potential reviewers, then please mention them here in this thread (without tagging them with an @). Feel free to use the [list of volunteers here](https://reviewers.joss.theoj.org/lookup) or the authors of the similar papers above.
