### 2024-04-26 #6506 @philbucher
Hi,
Thanks for the offer. Unfortunately this seems very far from my personal field of expertise, so I cannot do it.
I think the group of Thorsten Lutz at the IAG of the university of Stuttgart could be a good fit.
To my knowledge they are not using GitHub, so you would have to reach out via email
