### 2024-05-01 #6506 @paroomk
Hi @HaoZeke, sure. Happy to do it!

On Wed, May 1, 2024 at 6:00 AM Rohit Goswami ***@***.***>
wrote:

> hi @paroomk <https://github.com/paroomk> 👋 would you be interested in
> and available to review this JOSS submission? We carry out our
> checklist-driven reviews here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> If not, could you recommend any potential reviewers?
>
> P.S. Thanks @ashwinvis <https://github.com/ashwinvis> for catching the
> typo.
>
