+++ requirements.txt
@@ -10,4 +10,4 @@ setuptools==67.8.0
 sphinx_gallery==0.15.0
 tensorboard==2.15.0
 torch==1.13.0
-tqdm==4.65.0
+tqdm==4.66.4
