+++ requirements.txt
@@ -3,7 +3,7 @@ numpy==1.24.3
 plotly==5.17.0
 pyevtk==1.6.0
 pyFFTW==0.13.1
-pytest==7.4.2
+pytest==8.2.1
 scikit_learn==1.3.0
 scipy==1.10.1
 setuptools==67.8.0
