### 2024-05-27 1def184
:arrow_up: Bump scipy from 1.10.1 to 1.13.1
Bumps [scipy](https://github.com/scipy/scipy) from 1.10.1 to 1.13.1.
- [Release notes](https://github.com/scipy/scipy/releases)
- [Commits](https://github.com/scipy/scipy/compare/v1.10.1...v1.13.1)

---
updated-dependencies:
- dependency-name: scipy
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
