### 2024-06-04 #6506 @olivecha
Hi,

I would be available to review this.

> On Jun 3, 2024, at 5:09 PM, Rohit Goswami ***@***.***> wrote:
>
>
> hi @bhajay <https://github.com/bhajay>, @antviro <https://github.com/antviro> 👋 would you be interested in and available to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html <https://joss.readthedocs.io/en/latest/review_criteria.html>
> If not, could you recommend any potential reviewers?
>
