### 2024-06-04 #6838 @editorialbot
[REVIEW]: DRDMannTurb: A python package for scalable, data-driven synthetic turbulence
**Submitting author:** @alizma (<a href="http://orcid.org/0000-0003-3499-5555">Alexey Izmailov</a>)
**Repository:** https://github.com/METHODS-Group/DRDMannTurb
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.3
**Archive:** https://dx.doi.org/10.5281/zenodo.13922330
