### 2024-06-04 #6506 @antviro
> hi @bhajay, @antviro 👋 would you be interested in and available to review this JOSS submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
>
> If not, could you recommend any potential reviewers?

I see there are already two assigned reviewers, perhaps I am not needed anymore... But otherwise i can review it.
