### 2024-06-10 37e63c0
:arrow_up: Bump tensorboard from 2.15.0 to 2.17.0
Bumps [tensorboard](https://github.com/tensorflow/tensorboard) from 2.15.0 to 2.17.0.
- [Release notes](https://github.com/tensorflow/tensorboard/releases)
- [Changelog](https://github.com/tensorflow/tensorboard/blob/2.17.0/RELEASE.md)
- [Commits](https://github.com/tensorflow/tensorboard/compare/2.15.0...2.17.0)

---
updated-dependencies:
- dependency-name: tensorboard
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
