### 2024-07-04 METHODS-Group/DRDMannTurb#97 @paroomk
Include examples and documentation in the pre-compiled wheels for the package
I think the examples and documentation should be included in the pre-compiled wheels. This would make the package a lot more usable.

I had to clone the repo and follow the instructions under `Development Installation Instructions' to get access to the examples and documentation. Please let me know when you have fixed this.

cf: [](https://github.com/)openjournals/joss-reviews/issues/6838
