### 2024-07-04 #6838 @paroomk
> @paroomk @olivecha could you update me on the progress please?

Hi @HaoZeke, I am working on verifying the functionality now. I am hoping to finish this by next weekend (July 14).

I have also submitted an issue regarding the inclusion of examples and documentation in the pre-compiled wheels.
