### 2024-07-11 METHODS-Group/DRDMannTurb#98 @paroomk
[JOSS Review] Typo in references 
1. Please fix the typo in the last reference listed in the JOSS paper (DOI link).
2. Also, please provide the reference for HAWC2 mentioned in line 42 of the paper.

cf: [JOSS-6838](https://github.com/openjournals/joss-reviews/issues/6838)
