+++ paper/paper.bib
@@ -28,6 +28,13 @@ doi={10.5281/zenodo.7254149},
 version={1.0.0}
 }
 
+@software{hawc2:2007, 
+author={DTU Wind}, 
+year={2007}, 
+title={HaWC2: An Aeroelastic Code}, 
+url={https://www.hawc2.dk/}, 
+}
+
 @article{mann_spatial:1994,
 	author = {Mann, Jakob},
 	doi = {10.1017/S0022112094001886},
@@ -43,7 +50,7 @@ version={1.0.0}
 
 @article{mann_wind:1998,
 	author = {Mann, Jakob},
-	doi = {https://doi.org/10.1016/S0266-8920(97)00036-2},
+	doi = {10.1016/S0266-8920(97)00036-2},
 	issn = {0266-8920},
 	journal = {Probabilistic Engineering Mechanics},
 	number = {4},
