+++ paper/paper.md
@@ -32,7 +32,7 @@ Synthetic turbulence models (STMs) are used in wind engineering to generate real
 
 `DRDMannTurb` aims to create an easy-to-use framework to (1) fit one-point spectra from data using the DRD model introduced in [@Keith:2021] and (2) efficiently generate synthetic turbulence velocity fields to be used by scientists and engineers in downstream tasks. Existing methodologies for generating synthetic turbulence frequently incur a large computational overhead and lack the DRD model's flexibility to represent the diverse spectral properties of real-world observations, cf. [@Liew:2022]. `DRDMannTurb` addresses these two issues by introducing (1) a module for fitting neural network-based DRD models to observed one-point spectra data as well as (2) a module for efficiently generating synthetic turbulence boxes. Rather than generating turbulence for the entire box at once, which is a highly memory-intensive practice used in other software, `DRDMannTurb` uses a state-of-the-art domain decomposition approach to generate smaller sub-boxes sequentially.
 
-`DRDMannTurb` is completely written in Python, leveraging computationally powerful backend packages (`numpy`, `PyTorch`). Our implementation allows for easy GPU-portability using `cuda`. This is an additional advantage compared to previously developed software packages that have implemented the Mann model but do not provide the source code (e.g., HAWC2). Finally, `DRDMannTurb` is designed to be more general-purpose, allowing it to be applied to a broader range of scenarios and making it more accessible with clear documentation for a variety of tasks that researchers in this area are often interested in. 
+`DRDMannTurb` is completely written in Python, leveraging computationally powerful backend packages (`numpy`, `PyTorch`). Our implementation allows for easy GPU-portability using `cuda`. This is an additional advantage compared to previously developed software packages that have implemented the Mann model but do not provide the source code (e.g., HAWC2 [@hawc2:2007]). Finally, `DRDMannTurb` is designed to be more general-purpose, allowing it to be applied to a broader range of scenarios and making it more accessible with clear documentation for a variety of tasks that researchers in this area are often interested in. 
 
 # Results
 
