### 2024-07-12 METHODS-Group/DRDMannTurb#98 @alizma
Hi @paroomk, thank you for the comments. I updated the references and recompiled the article on our submission. Please confirm the changes so I can close the issue.
