### 2024-07-12 METHODS-Group/DRDMannTurb#98 @paroomk
Hi @alizma, references look good now. Thanks!
