### 2024-07-12 e640989
for #97, updated builds and releases with examples&docs
