+++ paper/paper.md
@@ -20,7 +20,7 @@ affiliations:
    index: 1
  - name: National Wind Technology Center, National Renewable Energy Laboratory, Golden, CO, 80401, USA
    index: 2
-date: 18 March 2024     
+date: 12 July 2024     
 bibliography: paper.bib
 ---
 
