+++ CITATION.cff
@@ -10,8 +10,8 @@ authors:
   - family-names: Keith
     given-names: Brendan
 title: "DRDMannTurb"
-version: 0.1.0
+version: 1.0.2
 identifiers:
   - type: doi
     value: 10.5281/zenodo.1234
-date-released: 2024-03-18
+date-released: 2024-07-12
