### 2024-07-12 METHODS-Group/DRDMannTurb#97 @alizma
Hi @paroomk, thank you for the suggestion. I updated the build system to include all of our examples and documentation in plain text. So, future users will have a plaintext version of examples they can run as well as the text of our documentation. Please see our new release ``1.0.2``, which is now on PyPi. The files can be seen [here](https://pypi.org/project/drdmannturb/#files).

I don't think the fully built documentation should be included in the wheels, as this would overwhelm the size of our code and make the installation size relatively large, which doesn't seem like a nice thing as a default, especially when our documentation page is always live.

Please confirm the changes so I can close the issue. Thanks again!
