### 2024-07-18 #6838 @paroomk
@HaoZeke Everything looks good to me now and I recommend this for publication in JOSS. Kudos to @alizma and co-authors!
