### 2024-07-29 METHODS-Group/DRDMannTurb#99 @olivecha
Unable to install development version on MacOS due to requirements
I tried installing the repository requirements using both `pip` and `conda`, and I run into troubles with both (see error logs).

I was able to install the development version by removing the specific package version requirements from the `requirement.txt` file.
The only conflict was to use `numpy<2.0`, this could either be added to the `requirement.txt` file, or the project could be made compatible with the latest `numpy` version (preferable). It seems only a single method is deprecated in `numpy` 2.0, from running the tests:

```
AttributeError: `np.string_` was removed in the NumPy 2.0 release. Use `np.bytes_` instead.. Did you mean: 'strings'?
```
[conda_error_log.txt](https://github.com/user-attachments/files/16416340/conda_error_log.txt)
[pip_error_log.txt](https://github.com/user-attachments/files/16416341/pip_error_log.txt)
