### 2024-07-29 METHODS-Group/DRDMannTurb#100 @olivecha
Test documentation and failing slow running tests
First, a small paragraph on how users can run the tests with `pytest` and what the machine requirements are (`cuda`), with a mention of the `--runslow` option would be expected in the `README`.

I was able to run the "fast" tests with success, but I encountered errors with the slow running tests.

- The `test_field_divergence` test fails due to missing `cuda` device (expected), but perhaps the test could be skipped in the absence of an available `cuda` device.
- The `test_nnparams_load_trained_TAUNET` and `test_nnparams_load_trained_CUSTOMMLP` fail because the object `OnePointSpectraDataGenerator` is instantiated without the `zref` argument:

```
    @pytest.mark.slow
    def test_nnparams_load_trained_CUSTOMMLP():
        pb = CalibrationProblem(
            nn_params=NNParameters(
                nlayers=2, hidden_layer_sizes=[10, 10], activations=[nn.GELU(), nn.ReLU()]
            ),
            prob_params=ProblemParameters(
                nepochs=2, eddy_lifetime=EddyLifetimeType.CUSTOMMLP
            ),
            loss_params=LossParameters(alpha_pen2=1.0, alpha_pen1=1.0e-5, beta_reg=2e-4),
            phys_params=PhysicalParameters(L=L, Gamma=Gamma, sigma=sigma, domain=domain),
            device=device,
        )

        k1_data_pts = domain
        DataPoints = [(k1, 1) for k1 in k1_data_pts]

>       Data = OnePointSpectraDataGenerator(data_points=DataPoints).Data
E       TypeError: OnePointSpectraDataGenerator.__init__() missing 1 required positional argument: 'zref'

test/io/test_model_params.py:123: TypeError
```

Please clarify if this is the expected behaviour.
