### 2024-07-29 METHODS-Group/DRDMannTurb#101 @olivecha
Missing community guidelines
Please add a paragraph in the `README` describing:

*"Clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"*
