### 2024-07-29 #6838 @olivecha
I've submitted three small issues about installation requirements, tests and missing community guidelines.

https://github.com/METHODS-Group/DRDMannTurb/issues/99
https://github.com/METHODS-Group/DRDMannTurb/issues/100
https://github.com/METHODS-Group/DRDMannTurb/issues/101
