+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,21 @@
+---
+name: Feature request
+about: Suggest a feature to help us improve
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**Problem you'd like to address**
+A clear and concise description of the problem you'd like to address. 
+
+- I'm always frustrated when [...];
+- `DRDMannTurb` would benefit from having [...];
+- and so on.
+
+**Propose your solution**
+A clear and concise description of your proposal.
+
+**Additional context**
+Add any other context about the feature request here.
