+++ README.md
@@ -1,8 +1,9 @@
-# Deep Rapid Distortion theory Mann Turbulence model
+# DRDMannTurb
 
 ![](https://github.com/METHODS-Group/DRDMannTurb/assets/74629347/604fcde9-41e1-4671-8c10-b1493cadfa88)
 
-
+DRDMannTurb (short for *Deep Rapid Distortion heory Mann Turbulence model*) is a data-driven framework
+for synthetic turbulence generation in Python.
 The code is based on the original work of Jacob Mann in [1994](https://doi.org/10.1017/S0022112094001886) and [1998](https://doi.org/10.1016/S0266-8920(97)00036-2)
 as well as in the deep-learning enhancement developed by Keith et al. in [this 2021 publication](https://arxiv.org/pdf/2107.11046.pdf). 
 
@@ -14,19 +15,6 @@ Pre-compiled wheels for the package are available via ``pip install drdmannturb`
 
 See the ``/examples/`` folder for baselines from the paper and for examples of the many functionalities of the package.
 
-## Development Installation Instructions 
-
-This package is provided as a locally ``pip``-installable package. It is recommended that Python 3.9 is used in your local environment. 
-
-A Conda spec-file is included, to install a functioning environment, run ``conda create --name ENV_NAME_HERE --file spec-file.txt``. Then run ``pip install -e .`` to load ``drdmannturb`` into your environment. 
-
-We also suggest installing the local ``pre-commit`` configuration by running ``pre-commit install`` in the root directory of this repository. 
-
-## Local Documentation Building Instructions 
-
-Docs are in the ``/docs/`` folder. Make sure the dependencies listed in ``./requirements-docs.txt`` are installed.
-
-Run ``make html`` to generate html pages in the ``/docs/build/html`` folder, which can be hosted locally with ``python -m http.server <PORT-NUMBER>``. 
 
 ## Citation 
 
@@ -42,6 +30,31 @@ version = {1.0.2},
 year = {2024}
 }
 ```
+
+## Contribute
+
+We always welcome new contributors! The best way to contribute to DRDMannTurb is through opening an issue, making a feature request, or creating a pull request directly.
+
+See also the below instructions for installing DRDMannTurb for development purposes.
+
+### Development Installation Instructions 
+
+This package is provided as a locally ``pip``-installable package. It is recommended that Python 3.9 is used in your local environment. 
+
+A Conda spec-file is included, to install a functioning environment, run ``conda create --name ENV_NAME_HERE --file spec-file.txt``. Then run ``pip install -e .`` to load ``drdmannturb`` into your environment. 
+
+We also suggest installing the local ``pre-commit`` configuration by running ``pre-commit install`` in the root directory of this repository. 
+
+### Local Documentation Building Instructions 
+
+Our documentation source lives in the ``/docs/`` folder. You shoudl ensure that the dependencies listed in ``./requirements-docs.txt`` are installed.
+
+Running ``make html`` will generate html pages in the ``/docs/build/html`` folder; these can be hosted locally with ``python -m http.server <PORT-NUMBER>``.
+
+## Questions?
+
+If you have any questions, the best way to receive help is by creating a thread in our Discussions or by contacting the authors (Alexey Izmailov, Matthew Meeker) directly. If your question pertains to a problem with the package itself, please open an Issue so that it can addressed.
+
 
 ## Acknowledgements
 
