### 2024-08-02 METHODS-Group/DRDMannTurb#101 @mjachi
Hi, I made a quick update to our README with a contributions and questions section and added templates for issues and feature requests. Please confirm the changes and we can close the issue.
