### 2024-08-02 METHODS-Group/DRDMannTurb#99 @mjachi
This is really an issue with Tensorboard dependency rather than with DRDMannTurb itself; see the below.
[Numpy 2.0 API compatibility issue #6869](https://github.com/tensorflow/tensorboard/issues/6869)
[Update usage of numpy to reflect numpy 2.0 changes #6871](https://github.com/tensorflow/tensorboard/pull/6871)

This has been merged and is on their nightly builds, so I believe all will be well with the next release.

Is it fine to temporarily fix `numpy<2.0`?
