### 2024-08-03 METHODS-Group/DRDMannTurb#99 @olivecha
Yes that would be fine, also the drdmanturb from the pre-compiled wheel does not pass tests at the moment due to changes in the `numpy` API, so you could temporarily add `numpy<2.0` in `setup.py`, then eventually replace it with a minimal Tensorboard version.

Also, you could specify in the development installation instructions that re-creating the exact conda environment is not mandatory (if not possible), and that you can just run `pip install -e . ` to use the development version, and dependencies will be resolved based on  "install requires" in `setup.py`.
