### 2024-08-04 d8cfd51
Numpy pin via setuptools; moved README animation to assets github hidden dir; platform specific spec files; README updates
