+++ conda_spec_files/macOS_arm64.txt
@@ -1,0 +1,214 @@
+# This file may be used to create an environment using:
+# $ conda create --name <env> --file <this file>
+# platform: osx-arm64
+@EXPLICIT
+https://repo.anaconda.com/pkgs/main/osx-arm64/libuv-1.48.0-h80987f9_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/blas-1.0-openblas.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/bzip2-1.0.8-h80987f9_6.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/c-ares-1.19.1-h80987f9_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/ca-certificates-2024.7.2-hca03da5_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/giflib-5.2.1-h80987f9_3.conda
+https://conda.anaconda.org/conda-forge/osx-arm64/icu-73.2-hc8870d7_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/jpeg-9e-h80987f9_2.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/libbrotlicommon-1.0.9-h80987f9_8.conda
+https://conda.anaconda.org/conda-forge/osx-arm64/libcxx-16.0.6-h4653b0c_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/libdeflate-1.17-h80987f9_1.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/libev-4.33-h1a28f6b_1.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/libffi-3.4.4-hca03da5_1.conda
+https://conda.anaconda.org/conda-forge/osx-arm64/libiconv-1.17-h0d3ecfb_2.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/libsodium-1.0.18-h1a28f6b_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/libwebp-base-1.3.2-h80987f9_0.conda
+https://conda.anaconda.org/conda-forge/osx-arm64/libzlib-1.2.13-hfb2fe0b_6.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/llvm-openmp-14.0.6-hc6e5704_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/ncurses-6.4-h313beb8_0.conda
+https://repo.anaconda.com/pkgs/main/noarch/pybind11-abi-4-hd3eb1b0_1.conda
+https://conda.anaconda.org/conda-forge/osx-arm64/python_abi-3.9-4_cp39.conda
+https://repo.anaconda.com/pkgs/main/noarch/tzdata-2024a-h04d1e81_0.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/xz-5.4.6-h80987f9_1.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/gmp-6.2.1-hc377ac9_3.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/gtest-1.14.0-h48ca7d4_1.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/lerc-3.0-hc377ac9_0.conda
+https://conda.anaconda.org/conda-forge/osx-arm64/libaec-1.1.2-h13dd4ca_1.conda
+https://repo.anaconda.com/pkgs/main/osx-arm64/libbrotlidec-1.0.9-h80987f9_8.conda
[…]
