### 2024-08-04 dd1ff98
Actions - Slow tests run on commit now; Minor updates to README
