+++ .github/workflows/run_tests.yml
@@ -21,7 +21,14 @@ jobs:
         run: | 
           pip install .
 
-      - name: Run tests
+      - name: Run fast tests
         run: | 
           pip install pytest
+          cd test
           python -m pytest 
+
+      - name: Run slow tests
+        run: | 
+          pip install pytest
+          cd test
+          python -m pytest --runslow
