### 2024-08-04 3b5dfaa
Update MM email; remove unused workflows (docs and main); specify tests and draft-pdf workflow execution conditions
