+++ .github/workflows/run_tests.yml
@@ -2,14 +2,25 @@ name: Tests
 
 on: [push, pull_request]
 
+concurrency:
+  group: ${{ github.workflow }}-${{ github.head_ref || github.run_id}}
+  cancel-in-progress: true
+
 jobs:
   run_tests:
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.9", "3.10", "3.11"]
+      fail-fast: false 
+
     steps: 
       - uses: actions/checkout@v4
-      - uses: actions/setup-python@v4
+
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v4
         with: 
-          python-version : '3.9'
+          python-version : ${{ matrix.python-version }}
           cache: 'pip'
 
       - name: Install dependencies
@@ -24,11 +35,8 @@ jobs:
       - name: Run fast tests
         run: | 
           pip install pytest
-          cd test
           python -m pytest 
 
       - name: Run slow tests
         run: | 
-          pip install pytest
-          cd test
           python -m pytest --runslow
