### 2024-08-04 METHODS-Group/DRDMannTurb#100 @mjachi
Hi. These have been fixed in #103 and the slow tests are now run on all push/pull-request to main.
