+++ .github/workflows/run_tests.yml
@@ -1,6 +1,13 @@
 name: Tests 
 
-on: [push, pull_request]
+on:
+  pull_request:
+    branches:
+      - main
+  push:
+    branches:
+      - main
+    tags: "*"
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.head_ref || github.run_id}}
@@ -8,11 +15,23 @@ concurrency:
 
 jobs:
   run_tests:
-    runs-on: ubuntu-latest
+    name: Python ${{ matrix.python-version }} - ${{ matrix.os }}
+    runs-on: ${{ matrix.os }}
+
     strategy:
-      matrix:
-        python-version: ["3.9", "3.10", "3.11"]
       fail-fast: false 
+      matrix:
+        python-version: 
+          - "3.9"
+          - "3.10" 
+          - "3.11"
+        os:
+          - ubuntu-latest
+        arch:
+          - x64
+        # group:
+        #   - Fast_Tests
+        #   - Slow_Tests
 
     steps: 
       - uses: actions/checkout@v4
