+++ .github/workflows/run_tests.yml
@@ -9,10 +9,6 @@ on:
       - main
     tags: "*"
 
-concurrency:
-  group: ${{ github.workflow }}-${{ github.head_ref || github.run_id}}
-  cancel-in-progress: true
-
 jobs:
   run_tests:
     name: Python ${{ matrix.python-version }} - ${{ matrix.os }}
