### 2024-08-04 METHODS-Group/DRDMannTurb#99 @mjachi
I can redo the wheels once things are more or less finalized.

#103 pins `numpy<2.0` in `setup.py`.

I should have caught this sooner, but our instructions in README were misleading -- the conda `spec-file.txt` was for Linux64 and therefore doesn't work on macOS (I myself am on an M2), but I didn't really see any clear indication that this was the issue in the error logs conda gave me (which, on my end, were about identical to what you provided in the OP) nor did we indicate that this would be an issue in the instructions. Just "unexpected error occurred".

I have included a more general `environment.yml` and moved all Conda environment exports under `conda_env_files`.

(As a sidenote, I am getting notifications about security risks with our dependencies from Dependabot; these are immediately resolved by bumping the versions up with the many PRs it has open currently. I will merge those before redoing the wheel so that tests pass)
