### 2024-08-06 METHODS-Group/DRDMannTurb#100 @olivecha
Could you just add that `--runslow` is a `pytest` option in the development installation instructions, so users who want to run the tests don't have to go read the test code to figure this out ?
