### 2024-08-07 6ddaa18
Minor README changes
