### 2024-08-09 #6838 @olivecha
@HaoZeke All issues were addressed by the authors, so I can recommend this for publication in the JOSS!
