### 2024-08-12 7f1e0b6
:arrow_up: Bump scikit-learn from 1.3.0 to 1.5.1
Bumps [scikit-learn](https://github.com/scikit-learn/scikit-learn) from 1.3.0 to 1.5.1.
- [Release notes](https://github.com/scikit-learn/scikit-learn/releases)
- [Commits](https://github.com/scikit-learn/scikit-learn/compare/1.3.0...1.5.1)

---
updated-dependencies:
- dependency-name: scikit-learn
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
