### 2024-08-12 d80f29d
:arrow_up: Bump matplotlib from 3.7.1 to 3.9.1.post1
Bumps [matplotlib](https://github.com/matplotlib/matplotlib) from 3.7.1 to 3.9.1.post1.
- [Release notes](https://github.com/matplotlib/matplotlib/releases)
- [Commits](https://github.com/matplotlib/matplotlib/compare/v3.7.1...v3.9.1.post1)

---
updated-dependencies:
- dependency-name: matplotlib
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
