+++ requirements.txt
@@ -1,4 +1,4 @@
-matplotlib==3.7.1
+matplotlib==3.9.1.post1
 numpy==1.26.4
 plotly==5.17.0
 pyevtk==1.6.0
