### 2024-08-12 ba86df0
:arrow_up: Bump plotly from 5.17.0 to 5.23.0
Bumps [plotly](https://github.com/plotly/plotly.py) from 5.17.0 to 5.23.0.
- [Release notes](https://github.com/plotly/plotly.py/releases)
- [Changelog](https://github.com/plotly/plotly.py/blob/master/CHANGELOG.md)
- [Commits](https://github.com/plotly/plotly.py/compare/v5.17.0...v5.23.0)

---
updated-dependencies:
- dependency-name: plotly
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
