### 2024-08-12 8a2be78
:arrow_up: Bump torch from 2.3.1 to 2.4.0
Bumps [torch](https://github.com/pytorch/pytorch) from 2.3.1 to 2.4.0.
- [Release notes](https://github.com/pytorch/pytorch/releases)
- [Changelog](https://github.com/pytorch/pytorch/blob/main/RELEASE.md)
- [Commits](https://github.com/pytorch/pytorch/compare/v2.3.1...v2.4.0)

---
updated-dependencies:
- dependency-name: torch
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
