+++ requirements.txt
@@ -9,5 +9,5 @@ scipy==1.10.1
 setuptools==70.0.0
 sphinx_gallery==0.15.0
 tensorboard==2.15.0
-torch==2.3.1
+torch==2.4.0
 tqdm==4.65.0
