### 2024-08-18 METHODS-Group/DRDMannTurb#115 @HaoZeke
[FEATURE] [JOSS] Consider using conda-lock
**Problem you'd like to address**
I noticed that #99 deals with version mismanagement between operating systems, perhaps `conda-lock` [1] might be of assistance here.

[1] https://github.com/conda/conda-lock

**Propose your solution**
Like any `lockfile` approach, `conda-lock` pins versions for later, the main difference is that new dependencies are explicitly resolved against the standard operating systems by default.

**Additional context**
Not really part of the JOSS review [2], but came up as part of it.

[2] https://github.com/openjournals/joss-reviews/issues/6838
