### 2024-08-18 METHODS-Group/DRDMannTurb#116 @HaoZeke
[JOSS] Missing reference
**Description**

**Reproduction/Minimal Working Example**
`@editorialbot check references`

**Expected behavior**
Should be clean. I believe the cannonical reference style here would be to either cite the version of the code closest to the publication most used (by the authors) from this list: https://www.hawc2.dk/hawc2-publications

Or (less fair TBH) to use the latest and the first publication references.

**Environment information**
N/A

**Additional context**
Final [JOSS tasks](https://github.com/openjournals/joss-reviews/issues/6838).
