### 2024-08-18 METHODS-Group/DRDMannTurb#117 @HaoZeke
[FEATURE] [JOSS] Consider using `pytest` groups
**Problem you'd like to address and suggestion**
Related to the discussion in #100, perhaps the design would benefit from separating the tests into classes which require specific devices or not.

**Additional context**
Another "not necessarily actionable" point [for JOSS](https://github.com/openjournals/joss-reviews/issues/6838), just one to keep in mind.
