### 2024-08-18 METHODS-Group/DRDMannTurb#118 @HaoZeke
[JOSS] Benchmark results
In particular, I note that the paper currently includes the line(s):

```
Once calibrated, a DRD model is no more expensive for simulating turbulence than a Mann model.
```

However, this should be borne out in a sample study / benchmark, and I'd assume there were reasonable limits to this (i.e. no unreasonably large number of parameters).

This could also be closed by just reframing the sentence a bit, perhaps something like, ".. has a good expense to cost ratio compared to a Mann model and is significantly more general" ?

Part of https://github.com/openjournals/joss-reviews/issues/6838
