### 2024-08-18 METHODS-Group/DRDMannTurb#119 @HaoZeke
[JOSS] Other minor paper comments
In no particular order (as of https://github.com/openjournals/joss-reviews/issues/6838#issuecomment-2295424005)...

## Line 14

```
to the Kolmogorov constant ...
```

For those not in the field this is not immediately physically relevant, consider adding few words here [fn1]?

## Line 25

```
simulator for generating turbulence boxes
```

Maybe a few words describing this as well (I know its covered quite nicely a few lines later though).

## Lines 39:40

```
DRDMannTurb is completely written in Python, leveraging computationally powerful backend packages (numpy, PyTorch).
```

It should be noted that these packages themselves utilize C / specialized frameworks for efficiently working on GPUs, otherwise it is a bit misleading, e.g. a pure PyPy package is also completely written in Python, without C altogether, which is not true of DRDMannTurb (or most other packages of course).

## Line 40

```
Our implementation allows for easy GPU-portability using cuda.
```

This seems overly restrictive, Pytorch can provide connections to other backends as well, mentioning cuda seems a bit disingenous.

## Line 42

```
 do not provide the source code
```

I haven't read the papers, but if they say anything like "will provide code on reasonable request" then this needs to be rephrased (though I agree / understand the sentiment this comes from).

## General comments

I am afraid for me, neural networks, and state of the art domain decomp without any other quantifiers / references feels rather overly general and weak, could these parts be elaborated on a bit?

Otherwise, it is definitely a nice piece of work :)

[fn1]: I liked https://spie.org/samples/FG02.pdf as a quick refresher...
