### 2024-08-19 c551bf6
:arrow_up: Bump tensorboard from 2.15.0 to 2.17.1
Bumps [tensorboard](https://github.com/tensorflow/tensorboard) from 2.15.0 to 2.17.1.
- [Release notes](https://github.com/tensorflow/tensorboard/releases)
- [Changelog](https://github.com/tensorflow/tensorboard/blob/master/RELEASE.md)
- [Commits](https://github.com/tensorflow/tensorboard/compare/2.15.0...2.17.1)

---
updated-dependencies:
- dependency-name: tensorboard
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
