+++ requirements.txt
@@ -6,7 +6,7 @@ pyFFTW==0.13.1
 pytest==7.4.2
 scikit_learn==1.3.0
 scipy==1.10.1
-setuptools==70.0.0
+setuptools==72.2.0
 sphinx_gallery==0.15.0
 tensorboard==2.15.0
 torch==2.3.1
