### 2024-08-23 METHODS-Group/DRDMannTurb#117 @mjachi
I'm leaning towards leaving the test suite as is since there are very few tests which require an available CUDA device. These are currently marked with `skipif`'s which are conditional on the availability of a CUDA device via PyTorch and set to name the absence of a usable CUDA device as the reason for these being skipped in the `pytest` output.

I think this approach is more convenient overall for now, but I do agree that it will probably make more sense to introduce separate classes whenever the test suite is expanded.
