### 2024-08-23 0e0c9d3
Reformat paper MD for readability
