### 2024-08-23 531ac75
Address issue 118 on benchmark results
