### 2024-08-23 465b2a2
Some paper edits
