### 2024-08-25 3db46f4
Update paper.md
Minor updates to the remainder of the paper
