### 2024-08-25 4d09743
Update paper.md
Figure captions
