+++ paper/paper.md
@@ -38,11 +38,11 @@ Synthetic turbulence models (STMs) are used in wind engineering to generate real
 
 The output of the fitting component of ``DRDMannTurb`` consists of two parts: the spectra fit by a DRD model and the learned eddy lifetime function. For example, in the case of a Kaimal spectra, the DRD spectra fit is more accurate than that of the Mann uniform shear model while providing an estimate of the physical parameters. 
 
-![Synthetic DRD Model Fit](synthetic_fit.png)
+![Synthetic DRD model fit.](synthetic_fit.png)
 
 After fitting to the spectra, the resulting models can also be used to generate 3D wind fields with spectra more closely resembling the same observations used in training.
 
-![Simulated Wind Tunnel](wind.png)
+![Simulated wind tunnel.](wind.png)
 
 For more detailed discussions of results, including a variety of utilities for interpolating and filtering noisy real-world data and generating wind turbulence, please see the [official examples](https://methods-group.github.io/DRDMannTurb/examples.html).
 
