### 2024-08-25 e7a50c1
Update paper.md
Figure captions
