### 2024-08-25 f16fe48
Update paper.md
Minor
