### 2024-08-26 6c6896e
:arrow_up: Bump setuptools from 70.0.0 to 73.0.1
Bumps [setuptools](https://github.com/pypa/setuptools) from 70.0.0 to 73.0.1.
- [Release notes](https://github.com/pypa/setuptools/releases)
- [Changelog](https://github.com/pypa/setuptools/blob/main/NEWS.rst)
- [Commits](https://github.com/pypa/setuptools/compare/v70.0.0...v73.0.1)

---
updated-dependencies:
- dependency-name: setuptools
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
