### 2024-08-26 METHODS-Group/DRDMannTurb#119 @mjachi
We've made the following the changes now.
- For Line 14, we added "influencing the global variance of the wind field"
- Line 25 has been replaced with "implementing an efficient method for synthetic turbulence generation based on
a domain decomposition technique."
- Line 39-40 has been rephrased to "`DRDMannTurb` is written in Python and leverages computationally powerful backend packages like `numpy` and `PyTorch`."
- Line 40 has been made more general: "The implementation makes DRD models easily portable to GPU and other backends via `PyTorch`."
- Line 42 has been changed to "but for which the source code may not be public or freely available"
