+++ paper/paper.bib
@@ -1,38 +1,34 @@
 @article{Keith:2021,
 	doi = {10.1063/5.0064394},
-  
 	url = {https://doi.org/10.1063%2F5.0064394},
-  
 	year = 2021,
 	month = {sep},
-  
 	publisher = {{AIP} Publishing},
-  
 	volume = {33},
-  
 	number = {9},
-  
 	author = {B. Keith and U. Khristenko and B. Wohlmuth},
-  
 	title = {Learning the structure of wind: A data-driven nonlocal turbulence model for the atmospheric boundary layer},
-  
 	journal = {Physics of Fluids}
 }
 
 @software{Liew:2022,
-author={{Liew, J.}},
-year = {2022}, 
-title = {{Mann.rs}},
-url={https://github.com/jaimeliew1/Mann.rs},
-doi={10.5281/zenodo.7254149}, 
-version={1.0.0}
+	author={{Liew, J.}},
+	year = {2022}, 
+	title = {{Mann.rs}},
+	url={https://github.com/jaimeliew1/Mann.rs},
+	doi={10.5281/zenodo.7254149}, 
+	version={1.0.0}
 }
 
-@software{hawc2:2007, 
-author={DTU Wind}, 
-year={2007}, 
-title={HaWC2: An Aeroelastic Code}, 
-url={https://www.hawc2.dk/}, 
+@article{
+	author={Madsen, H. A. and Larsen, T. J. and Pirrung, G. R. and Li, A. and Zahle, F.},
+	title={Implementation of the blade element momentum model on a polar grid and its aeroelastic load impact},
+	journal = {Wind Energy Science},
+	doi = {10.5194/wes-5-1-2020},
+	pages = {1--27},
+	year = {2020},
+	volume = {5},
+	url = {https://wes.copernicus.org/articles/5/1/2020/},
 }
 
 @article{mann_spatial:1994,
@@ -45,8 +41,6 @@ url={https://www.hawc2.dk/},
 	volume = {273},
 	year = {1994},
 }
-
-
 
 @article{mann_wind:1998,
 	author = {Mann, Jakob},
