### 2024-08-26 METHODS-Group/DRDMannTurb#116 @mjachi
I've just updated the citation. It seems that the HAWC2 authors/DTU affiliates just reference the HAWC2 user manual in their publications, except the user manual is not published. We opted to cite their [most recent journal publication](https://wes.copernicus.org/articles/5/1/2020/#section2) they have listed there.
