+++ paper/paper.bib
@@ -20,9 +20,9 @@
 	version={1.0.0}
 }
 
-@article{
-	author={Madsen, H. A. and Larsen, T. J. and Pirrung, G. R. and Li, A. and Zahle, F.},
-	title={Implementation of the blade element momentum model on a polar grid and its aeroelastic load impact},
+@article{wes-5-1-2020,
+	author = {Madsen, H. A. and Larsen, T. J. and Pirrung, G. R. and Li, A. and Zahle, F.},
+	title = {Implementation of the blade element momentum model on a polar grid and its aeroelastic load impact},
 	journal = {Wind Energy Science},
 	doi = {10.5194/wes-5-1-2020},
 	pages = {1--27},
