### 2024-08-29 560e3f5
Base development environment lock file for (currently ARM64 macos only)
