+++ conda_env_files/environment.yml
@@ -1,311 +1,25 @@
-name: DRDMT
+name: drdmannturb_env
 channels:
   - conda-forge
   - defaults
 dependencies:
-  - abseil-cpp=20230802.0
-  - absl-py=2.1.0
-  - aiohttp=3.9.5
-  - aiosignal=1.2.0
-  - alabaster=0.7.16
-  - appnope=0.1.3
-  - asttokens=2.4.0
-  - async-timeout=4.0.3
-  - attrs=23.1.0
-  - babel=2.11.0
-  - backcall=0.2.0
-  - backports=1.1
-  - backports.functools_lru_cache=1.6.5
-  - black=24.4.2
-  - blas=1.0
-  - blinker=1.6.2
-  - blosc=1.21.5
-  - bottleneck=1.3.7
-  - brotli=1.0.9
-  - brotli-bin=1.0.9
-  - brotli-python=1.0.9
-  - brotlipy=0.7.0
-  - bzip2=1.0.8
-  - c-ares=1.19.1
-  - ca-certificates=2024.7.2
-  - cachetools=5.3.3
-  - certifi=2024.7.4
-  - cffi=1.16.0
-  - cftime=1.6.3
-  - charset-normalizer=3.3.2
-  - click=8.1.7
-  - colorama=0.4.6
-  - comm=0.2.1
-  - contourpy=1.2.0
-  - coverage=7.2.2
-  - cryptography=42.0.5
-  - cycler=0.11.0
-  - debugpy=1.6.7
-  - decorator=5.1.1
-  - docutils=0.18.1
-  - entrypoints=0.4
-  - exceptiongroup=1.2.0
-  - executing=1.2.0
-  - fftw=3.3.10
-  - filelock=3.13.1
-  - fonttools=4.51.0
-  - freetype=2.12.1
-  - frozenlist=1.4.0
-  - fsspec=2024.3.1
-  - giflib=5.2.1
-  - gmp=6.2.1
-  - gmpy2=2.1.2
-  - google-auth=2.29.0
-  - google-auth-oauthlib=0.5.2
-  - grpc-cpp=1.48.2
-  - grpcio=1.48.2
-  - gtest=1.14.0
-  - hdf4=4.2.15
-  - hdf5=1.14.3
-  - icu=73.2
-  - idna=3.7
-  - imagesize=1.4.1
-  - importlib-metadata=7.0.1
-  - importlib_metadata=7.0.1
-  - importlib_resources=6.4.0
-  - iniconfig=1.1.1
-  - ipykernel=6.28.0
-  - ipython=8.16.1
-  - isort=5.13.2
-  - jedi=0.19.1
-  - jinja2=3.1.4
-  - jpeg=9e
-  - jupyter_client=8.6.0
-  - jupyter_core=5.7.2
-  - kiwisolver=1.4.4
-  - krb5=1.20.1
-  - lcms2=2.12
-  - lerc=3.0
-  - libaec=1.1.2
-  - libbrotlicommon=1.0.9
-  - libbrotlidec=1.0.9
-  - libbrotlienc=1.0.9
-  - libcurl=8.7.1
-  - libcxx=16.0.6
-  - libdeflate=1.17
-  - libedit=3.1.20230828
-  - libev=4.33
-  - libffi=3.4.4
-  - libgfortran=5.0.0
-  - libgfortran5=13.2.0
-  - libiconv=1.17
-  - libnetcdf=4.9.2
-  - libnghttp2=1.57.0
-  - libopenblas=0.3.21
-  - libpng=1.6.39
-  - libprotobuf=3.20.3
-  - libsodium=1.0.18
-  - libsqlite=3.46.0
[…]
