### 2024-08-29 9cc9874
Conda locks for all platforms with testing and docs dependencies, updated README
