### 2024-09-09 51ae4b9
:arrow_up: Bump setuptools from 70.0.0 to 74.1.2
Bumps [setuptools](https://github.com/pypa/setuptools) from 70.0.0 to 74.1.2.
- [Release notes](https://github.com/pypa/setuptools/releases)
- [Changelog](https://github.com/pypa/setuptools/blob/main/NEWS.rst)
- [Commits](https://github.com/pypa/setuptools/compare/v70.0.0...v74.1.2)

---
updated-dependencies:
- dependency-name: setuptools
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
