### 2024-09-16 217d8ae
:arrow_up: Bump setuptools from 70.0.0 to 75.0.0
Bumps [setuptools](https://github.com/pypa/setuptools) from 70.0.0 to 75.0.0.
- [Release notes](https://github.com/pypa/setuptools/releases)
- [Changelog](https://github.com/pypa/setuptools/blob/main/NEWS.rst)
- [Commits](https://github.com/pypa/setuptools/compare/v70.0.0...v75.0.0)

---
updated-dependencies:
- dependency-name: setuptools
  dependency-type: direct:production
  update-type: version-update:semver-major
...

Signed-off-by: dependabot[bot] <support@github.com>
