### 2024-09-25 METHODS-Group/DRDMannTurb#115 @HaoZeke
Excellent, closing

> Thank you for the suggestion! I added conda-lock in #127 and have updated the README. Let us know if you have any other thoughts on this or happen to catch any issues.
