### 2024-09-25 #6838 @HaoZeke
Excellent, we're almost at the finish line, @alizma could you make a final release for JOSS and put the DOI here following the instructions in this comment? https://github.com/openjournals/joss-reviews/issues/6838#issuecomment-2278131165
