### 2024-10-12 #6838 @alizma
Hi @HaoZeke, apologies for the delay, life got a bit busy. I made a new release (v1.0.3 that syncs all the changes from the reviews and uploaded to a Zenodo archive, please find it [here](https://zenodo.org/records/13922330). Please let me know if there's anything else to complete on our end. Thank you for your help and feedback!
