### 2024-10-14 a4a8e57
:arrow_up: Bump sphinx-gallery from 0.15.0 to 0.18.0
Bumps [sphinx-gallery](https://github.com/sphinx-gallery/sphinx-gallery) from 0.15.0 to 0.18.0.
- [Release notes](https://github.com/sphinx-gallery/sphinx-gallery/releases)
- [Changelog](https://github.com/sphinx-gallery/sphinx-gallery/blob/master/.github_changelog_generator)
- [Commits](https://github.com/sphinx-gallery/sphinx-gallery/compare/v0.15.0...v0.18.0)

---
updated-dependencies:
- dependency-name: sphinx-gallery
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <supp
[…]
