+++ requirements.txt
@@ -7,7 +7,7 @@ pytest==7.4.2
 scikit_learn==1.3.0
 scipy==1.10.1
 setuptools==70.0.0
-sphinx_gallery==0.15.0
+sphinx_gallery==0.18.0
 tensorboard==2.15.0
 torch==2.3.1
 tqdm==4.65.0
