### 2024-10-18 #6838 @crvernon
:wave: @alizma - I need you to address the following before I accept this submission for publication:

In your archive:

- [ ] The archive name (including capitalization) and author list (including order) in your Zenodo archive should match what is in your paper exactly.  Also, please ensure that the license you list in your Zenodo archive matches the one you use in your GitHub repository.  Please edit the metadata in your Zenodo archive to fix these things.  A new release is not required.

In your paper:

- [ ] LINE 26: The citation format for this should be changed from "...described in (Keith et al., 2021)." to "...described in Keith et al. (2021)."
- [ ] LINE 29:  Same as the above comment about the citation format.
- [ ] LINE 40: "numpy" should be written in this case as "NumPy"
- [ ] LINE 70:  Is there something missing in this statement?  "The U.S. Government retains and the publisher, by accepting the article..."

Once you adderss the above let me know and I'll accept this for publication.  Thanks!
