### 2024-10-23 2d0919b
Paper updates
