### 2024-10-23 #6838 @alizma
Hi @crvernon please see the updated metadata on the [Zenodo archive here](https://zenodo.org/records/13922330). I updated the title, description and license. Please let us know if there is anything else. Thank you!
